Use Git to print diff in style check
The script apply_style now always runs in 'fix' mode, i.e. it will
traverse the directory tree and reformat files that do not conform to
the style.

For use on the build server, the -d/--diff argument may be used. The
script will then call out to Git in order to generate the diff of the
changes. If there are changes, the script will print the diff and exit
with an error.