Word2vec clients shouldn't need to have the same vocab (#1685)
Modifying the word2vec client so that clients don't need to have the same vocabulary. Changes made: Gradient updates now translated from the sender's vocab to the receiver's.
Clients need a method for sending their vocab to other clients on request.
When weights are synchronised in synchronised mode the weights also need to be translated by each client. Also moving digest.hpp to core from ledger so that it can be used in ML example